prompts/src/requirements.txt
python-liquid==2.1.0
